+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support `--page-size` to determine number of items shown in interactive mode
 - Support for V2 NuGet feeds
 - Support for `nuget.config` package sources
+- Support for `--no-progress` to hide progress bar while fetching package information
 
 ### Changed
 - Handling when no packages match provided filters (`--include` or `--exclude`)