+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.1] / 2024-10-29
 - fix: Xml declaration not being saved correctly
 
 ## [0.3.0] / 2024-02-25
@@ -44,7 +46,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] / 2023-11-10
 - Initial release
 
-[Unreleased]: https://github.com/vipentti/dotnet-check-updates/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/vipentti/dotnet-check-updates/compare/0.3.1...HEAD
+[0.3.1]: https://github.com/vipentti/dotnet-check-updates/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/vipentti/dotnet-check-updates/compare/0.2.4...0.3.0
 [0.2.4]: https://github.com/vipentti/dotnet-check-updates/compare/0.2.3...0.2.4
 [0.2.3]: https://github.com/vipentti/dotnet-check-updates/compare/0.2.2...0.2.3